+++ app/src/main/java/com/snb/repeater/app/domain/db/DBAbstract.java
@@ -3,6 +3,7 @@ package com.snb.repeater.app.domain.db;
 import android.arch.persistence.room.Database;
 import android.arch.persistence.room.RoomDatabase;
 
+import com.snb.repeater.app.domain.model.DB;
 
 /**
  * Created by
@@ -12,6 +13,5 @@ import android.arch.persistence.room.RoomDatabase;
 @Database(entities = {DB.class}, version = 1, exportSchema = false)
 public abstract class DBAbstract extends RoomDatabase {
 
-        public abstract DBDao getDBDao(); //return all metods
-
+    public abstract DBDao getDBDao(); //return all metods
 }
