+++ app/src/main/java/com/snb/repeater/app/domain/db/DBDao.java
@@ -5,7 +5,7 @@ import android.arch.persistence.room.Delete;
 import android.arch.persistence.room.Insert;
 import android.arch.persistence.room.Query;
 
-
+import com.snb.repeater.app.domain.model.DB;
 
 import java.util.List;
 
@@ -16,16 +16,15 @@ import java.util.List;
  */
 @Dao
 public interface DBDao {
-        // Add question to DB
-        @Insert
-        void insert(DB db);
+    // Add question to DB
+    @Insert
+    void insert(DB db);
 
-        // Delete question from DB
-        @Delete
-        void delete(DB db);
+    // Delete question from DB
+    @Delete
+    void delete(DB db);
 
-        // Get questions from DB
-        @Query("SELECT * FROM db ")
-        List<DB> getAllData();
-
+    // Get questions from DB
+    @Query("SELECT * FROM db ")
+    List<DB> getAllData();
 }
